Week-8/app/requirements.txt
flask
python-dotenv
requests
python-jose
python3-saml
authlib

